Apply #1 - Component 적용 컴포넌트 적용
