+++ src/main/java/org/example/logic/BubbleSort.java
@@ -1,8 +1,12 @@
 package org.example.logic;
 
+import org.springframework.stereotype.Component;
+
 import java.util.ArrayList;
 import java.util.List;
 
+
+@Component
 public class BubbleSort <T extends Comparable<T>> implements Sort<T>{
 
     @Override
