+++ src/main/java/org/example/logic/JavaSort.java
@@ -1,9 +1,12 @@
 package org.example.logic;
 
+import org.springframework.stereotype.Component;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@Component
 public class JavaSort <T extends Comparable<T>> implements Sort<T>{
 
     @Override
