--- src/main/java/service/SortService.java
@@ -1,19 +1,0 @@
-package service;
-
-import org.example.logic.Sort;
-
-import java.util.List;
-
-public class SortService {
-
-    private final Sort<String> sort;
-    //생성자 주입 방식 1번
-    public SortService(Sort<String> sort) {
-        this.sort = sort;
-        System.out.println("[구현체] :" +sort.getClass().getName());
-    }
-
-    public List<String> doSort(List<String> list) {
-        return sort.sort(list);
-    }
-}
